+++ roocs_utils/parameter/time_components_parameter.py
@@ -1,66 +1,53 @@
-import datetime
-
-from dateutil import parser as date_parser
-
 from roocs_utils.exceptions import InvalidParameterValue
 from roocs_utils.parameter.base_parameter import _BaseParameter
+from roocs_utils.parameter.param_utils import time_components, string_to_dict
 
 
 class TimeComponentsParameter(_BaseParameter):
     """
     Class for time components parameter used in subsetting operation.
 
-*************
+    The Time Components are any, or none of:
+      - year: [list of years]
+      - month: [list of months]
+      - day: [list of days]
+      - hour: [list of hours]
+      - minute: [list of minutes]
+      - second: [list of seconds]
 
+    `month` is special: you can use either strings or values:
+       "feb", "mar" == 2, 3 == "02,03"
 
-NEED TO DOCUMENT
-
-
-****
-
-
-    Validates the times input and parses the values into isoformat.
-
+    Validates the times input and parses the them into a dictionary.
     """
 
- #   parse_method = "_parse_range"
+    allowed_input_types = [dict, str, time_components, type(None)]
 
-    def _validate(self):
+    def _parse(self):
         try:
-            self._parse_times()
-
-        except (date_parser._parser.ParserError, TypeError):
-            raise InvalidParameterValue("Unable to parse the time values entered")
-
-    def _parse_times(self):
-        # should this default to the start and end time of the data?
-        start, end = self._result
-
-        if start is not None:
-            start = date_parser.parse(
-                start, default=datetime.datetime(datetime.MINYEAR, 1, 1)
-            ).isoformat()
-        if end is not None:
-            end = date_parser.parse(
-                end, default=datetime.datetime(datetime.MAXYEAR, 12, 30)
-            ).isoformat()
-
-        return start, end
-
-    @property
-    def tuple(self):
-        """Returns a tuple of the time values"""
-        if self._parse_times() is not (None, None):
-            return self._parse_times()
+            if self.input in (None, ""):
+                return None
+            elif isinstance(self.input, time_components):
+                return self.input.value
+            elif isinstance(self.input, str):
+                time_comp_dict = string_to_dict(self.input, splitters=("|", ":", ","))
+                return time_components(**time_comp_dict).value
+            else:  # Must be a dict to get here
+                return time_components(**self.input).value
+        except Exception:
+            raise InvalidParameterValue(
+                f"Cannot create TimeComponentsParameter " f"from: {self.input}"
+            )
 
     def asdict(self):
-        """Returns a dictionary of the time values"""
-        if self.tuple is not None:
-            return {"start_time": self.tuple[0], "end_time": self.tuple[1]}
+        # Just return the value, either a dict or None
+        return self.value
 
     def __str__(self):
-        return (
-            f"Time period to subset over"
-            f"\n start time: {self.tuple[0]}"
-            f"\n end time: {self.tuple[1]}"
-        )
+        if self.value is None:
+            return "No time components specified"
+
+        resp = "Time components to select:"
+        for key, value in self.value.items():
+            resp += f"\n    {key} => {value}"
+        return resp