+++ roocs_utils/parameter/time_parameter.py
@@ -2,8 +2,13 @@ import datetime
 
 from roocs_utils.exceptions import InvalidParameterValue
 from roocs_utils.parameter.base_parameter import _BaseIntervalOrSeriesParameter
-from roocs_utils.parameter.param_utils import (parse_range, parse_sequence,
-                            parse_datetime, interval, series)
+from roocs_utils.parameter.param_utils import (
+    parse_range,
+    parse_sequence,
+    parse_datetime,
+    interval,
+    series,
+)
 
 
 class TimeParameter(_BaseIntervalOrSeriesParameter):
@@ -29,11 +34,15 @@ class TimeParameter(_BaseIntervalOrSeriesParameter):
 
         try:
             if start is not None:
-                start = parse_datetime(start, default=datetime.datetime(datetime.MINYEAR, 1, 1))
+                start = parse_datetime(
+                    start, default=datetime.datetime(datetime.MINYEAR, 1, 1)
+                )
             if end is not None:
-                end = parse_datetime(end, default=datetime.datetime(datetime.MAXYEAR, 12, 30))
+                end = parse_datetime(
+                    end, default=datetime.datetime(datetime.MAXYEAR, 12, 30)
+                )
 
-        except:
+        except Exception:
             raise InvalidParameterValue("Unable to parse the time values entered")
 
         # Set as None if no start or end, otherwise set as tuple
@@ -47,7 +56,7 @@ class TimeParameter(_BaseIntervalOrSeriesParameter):
     def _parse_as_series(self):
         try:
             value = [parse_datetime(tm) for tm in self.input.value]
-        except:
+        except Exception:
             raise InvalidParameterValue("Unable to parse the time values entered")
 
         return value