+++ tests/test_parameter/test_time_parameter.py
@@ -4,13 +4,18 @@ import pytest
 
 from roocs_utils.exceptions import InvalidParameterValue
 from roocs_utils.parameter.time_parameter import TimeParameter
-from roocs_utils.parameter.param_utils import (time_interval, time_series,
-                                               time_components)
+from roocs_utils.parameter.param_utils import (
+    time_interval,
+    time_series,
+    time_components,
+)
 
 
-type_err = ("Input type of <{}> not allowed. Must be one of: [<class "
-            "'roocs_utils.parameter.param_utils.Interval'>, "
-            "<class 'roocs_utils.parameter.param_utils.Series'>, <class 'NoneType'>]")
+type_err = (
+    "Input type of <{}> not allowed. Must be one of: [<class "
+    "'roocs_utils.parameter.param_utils.Interval'>, "
+    "<class 'roocs_utils.parameter.param_utils.Series'>, <class 'NoneType'>]"
+)
 
 
 def test__str__():
@@ -34,10 +39,7 @@ def test_raw():
 def test_validate_error_len_1_tuple():
     with pytest.raises(InvalidParameterValue) as exc:
         time_interval(("2085-01-01T12:00:00Z",))
-    assert (
-        str(exc.value)
-        == "Interval should be a range. Expected 2 values, received 1"
-    )
+    assert str(exc.value) == "Interval should be a range. Expected 2 values, received 1"
 
 
 # should datetime objects be allowed?
@@ -91,20 +93,20 @@ def test_as_dict():
 def test_slash_none():
     time = time_interval("/")
     parameter = TimeParameter(time)
-    assert parameter.value == None
+    assert parameter.value is None
     assert parameter.asdict() == {"start_time": None, "end_time": None}
 
 
 def test_none():
     time = None
     parameter = TimeParameter(time)
-    assert parameter.value == None
+    assert parameter.value is None
 
 
 def test_empty_string():
     time = time_interval("")
     parameter = TimeParameter(time)
-    assert parameter.value == None
+    assert parameter.value is None
 
     with pytest.raises(InvalidParameterValue) as exc:
         TimeParameter("")